+++ R/plotHeatmap-methods.R
@@ -1,10 +1,19 @@
 #' @name plotHeatmap
 #' @inherit acidplots::plotHeatmap description return title
-#' @note Updated 2020-01-20.
+#' @note Updated 2020-09-23.
 #'
 #' @inheritParams acidroxygen::params
 #' @inheritParams params
 #' @param ... Additional arguments.
+#'
+#' @examples
+#' data(fgsea)
+#' plotHeatmap(
+#'     object = fgsea,
+#'     contrast = "condition_B_vs_A",
+#'     collection = "h",
+#'     set = "HALLMARK_P53_PATHWAY"
+#' )
 NULL
 
 
@@ -18,59 +27,44 @@ NULL
 
 
 
-## Updated 2019-11-19.
+## Updated 2020-09-23.
 `plotHeatmap,FGSEAList` <-  # nolint
     function(
         object,
-        DESeqAnalysis,  # nolint
         contrast,
         contrastSamples = TRUE,
         collection,
         set,
-        leadingEdge = TRUE,
+        leadingEdge = FALSE,
         ...
     ) {
         validObject(object)
-        validObject(DESeqAnalysis)
         assert(
-            is(DESeqAnalysis, "DESeqAnalysis"),
-            identical(
-                x = names(object[[1L]]),
-                y = contrastNames(DESeqAnalysis)
-            ),
-            isScalar(contrast),
+            isString(contrast),
             isFlag(contrastSamples),
-            isScalar(collection),
+            isString(collection),
             isString(set),
             isFlag(leadingEdge)
         )
-        ## Match collection to name, if necessary.
-        if (!isString(collection)) {
-            collection <- names(object)[[collection]]
-        }
-        ## Match contrast to name, if necessary.
-        if (!isString(contrast)) {
-            contrast <- names(object[[collection]])[[contrast]]
-        }
         ## Map the genes we want to plot to the DESeq data.
         if (isTRUE(leadingEdge)) {
-            genes <- .leadingEdge(
+            genes <- leadingEdge(
                 object = object,
                 contrast = contrast,
                 collection = collection,
                 set = set
             )
         } else {
-            ## Locate the GMT file used to run GSEA.
-            gmt <- import(file = metadata(object)[["gmtFiles"]][[collection]])
-            genes <- gmt[[set]]
+            genes <- NULL
         }
         ## Plot the log counts from DESeqTransform object.
-        dt <- as(DESeqAnalysis, "DESeqTransform")
-        rownames <- mapGenesToRownames(object = dt, genes = genes)
-        dt <- dt[rownames, , drop = FALSE]
+        genes <- geneSet(object, collection = collection, set = set)
+        rownames <- .matchGenesToIDs(object, set = set, genes = genes)
+        deseq <- `DESeqAnalysis,FGSEAList`(object)
+        dt <- as(deseq, "DESeqTransform")
+        dt <- dt[rownames, ]
         if (isTRUE(contrastSamples)) {
-            colnames <- contrastSamples(DESeqAnalysis, i = contrast)
+            colnames <- contrastSamples(deseq, i = contrast)
             dt <- dt[, colnames, drop = FALSE]
         }
         args <- list(